Support glyph indices beyond 65535 throughout font subsystem
Extended font subsystem from 16-bit to 32-bit glyph indices to support
large fonts (massive CJK fonts, emoji fonts with many variants).

Core Changes:
- Changed FontFile.CharacterToGlyphIndex from Dictionary<int, ushort> to Dictionary<int, uint>
- Updated GetGlyphIndex() return type from ushort? to uint?
- Updated ShapedGlyph.GlyphId from ushort to uint

Implementation Details:
1. **FontFile.cs**: Updated all glyph index handling to use uint
   - GetAdvanceWidth, GetKerning, GetGlyphData methods updated
   - Added bounds checking for legacy kerning (still uses ushort pairs)
   - Documented that kern table doesn't support glyphs > 65535

2. **CmapTableParser.cs**: Removed 0xFFFF limitation in ParseFormat12
   - Now stores full range of glyph indices from format 12 tables
   - Removed TODO about supporting glyphs beyond 65535

3. **FontSubsetter.cs**: Updated glyph mapping
   - Changed BuildGlyphMapping to use Dictionary<uint, ushort>
   - Input font can have uint indices, output subset uses ushort (always < 65536 glyphs)
   - Updated all iteration and array indexing with proper casts

4. **OpenTypeShaper.cs**: Updated text shaping
   - Changed ShapedGlyph.GlyphId to uint
   - Added bounds checking for GSUB/GPOS operations (use 16-bit indices)
   - Glyphs > 65535 skip ligatures, substitutions, and positioning
   - This is correct: OpenType GSUB/GPOS tables use 16-bit glyph IDs

5. **TrueTypeFontSerializer.cs**: Added explicit casts
   - Format 4 cmap tables use 16-bit indices
   - Safe for subset fonts which always have < 65536 glyphs

Compatibility Notes:
- Legacy tables (kern, GSUB, GPOS) still use 16-bit glyph indices per spec
- Fonts with > 65535 glyphs won't have kerning/ligatures for high-index glyphs
- This is acceptable: such massive fonts are extremely rare

Files modified:
- src/Folly.Fonts/Models/FontFile.cs
- src/Folly.Fonts/Tables/CmapTableParser.cs
- src/Folly.Fonts/FontSubsetter.cs
- src/Folly.Fonts/OpenType/OpenTypeShaper.cs
- src/Folly.Fonts/TrueTypeFontSerializer.cs

+++ src/Folly.Fonts/TrueTypeFontSerializer.cs
@@ -640,7 +640,8 @@ public class TrueTypeFontSerializer
         var segments = new List<(ushort startCode, ushort endCode, short idDelta)>();
         int rangeStart = sortedChars[0].Key;
         int rangeEnd = rangeStart;
-        ushort firstGlyphInRange = sortedChars[0].Value;
+        // Format 4 cmap uses 16-bit glyph indices - cast from uint
+        ushort firstGlyphInRange = (ushort)sortedChars[0].Value;
 
         for (int i = 1; i < sortedChars.Count; i++)
         {
@@ -659,7 +660,7 @@ public class TrueTypeFontSerializer
 
                 rangeStart = currentChar;
                 rangeEnd = currentChar;
-                firstGlyphInRange = sortedChars[i].Value;
+                firstGlyphInRange = (ushort)sortedChars[i].Value;
             }
         }
